Adjust imports and namespaces after refactor
