+++ GraphQL.sln
@@ -9,6 +9,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "GraphQL.Tests", "src\GraphQ
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "GraphQL.GraphiQL", "src\GraphQL.GraphiQL\GraphQL.GraphiQL.csproj", "{EE557888-1CD2-4E87-BD87-06834C988B23}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "GraphQL.StarWars", "GraphQL.StarWars\GraphQL.StarWars.csproj", "{3255C0DC-9F24-4423-8BA3-3B979D0E90E2}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -27,6 +29,10 @@ Global
 		{EE557888-1CD2-4E87-BD87-06834C988B23}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{EE557888-1CD2-4E87-BD87-06834C988B23}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{EE557888-1CD2-4E87-BD87-06834C988B23}.Release|Any CPU.Build.0 = Release|Any CPU
+		{3255C0DC-9F24-4423-8BA3-3B979D0E90E2}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{3255C0DC-9F24-4423-8BA3-3B979D0E90E2}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{3255C0DC-9F24-4423-8BA3-3B979D0E90E2}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{3255C0DC-9F24-4423-8BA3-3B979D0E90E2}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
