+++ src/GraphQL.GraphiQL/GraphQL.GraphiQL.csproj
@@ -199,9 +199,9 @@
     <Content Include="packages.config" />
   </ItemGroup>
   <ItemGroup>
-    <ProjectReference Include="..\GraphQL.Tests\GraphQL.Tests.csproj">
-      <Project>{084e1e35-14bd-414b-97b6-1a6469cd1893}</Project>
-      <Name>GraphQL.Tests</Name>
+    <ProjectReference Include="..\..\GraphQL.StarWars\GraphQL.StarWars.csproj">
+      <Project>{3255c0dc-9f24-4423-8ba3-3b979d0e90e2}</Project>
+      <Name>GraphQL.StarWars</Name>
     </ProjectReference>
     <ProjectReference Include="..\GraphQL\GraphQL.csproj">
       <Project>{c0e08b04-8a0d-4196-967c-8850825041b0}</Project>
