+++ src/GraphQL.GraphiQL/SimpleContainerDependencyResolver.cs
@@ -2,7 +2,8 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Web.Http.Dependencies;
-using GraphQL.Tests;
+using GraphQL.StarWars;
+using GraphQL.StarWars.IoC;
 
 namespace GraphQL.GraphiQL
 {
