+++ src/GraphQL.Tests/StarWars/StarWarsBasicQueryTests.cs
@@ -1,4 +1,4 @@
-namespace GraphQL.Tests
+namespace GraphQL.Tests.StarWars
 {
     public class StarWarsBasicQueryTests : StarWarsTestBase
     {
